Gui don't deselect files on model change
If anything in the model of the gui changes, the view is
updated accordingly. However, selected files where marked
"un-selected" because the look up was based on filename, whereas
the model used full path names. Now the lookup is on full path
names once, and the apppropriate files are selected.

this fixes #1